requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
colorama==0.4.6
distro==1.9.0
Django==5.0.6
django-bootstrap5==24.2
groq==0.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
